refactor: remove inline editable directive, editor adapter interface, and related services
