+++ ui/src/app/components/core/document-sections/document-sections.component.html
@@ -1,0 +1,17 @@
+<div class="document-sections w-full h-full flex flex-col">
+  <div class="sidebar-header p-4 border-b flex-shrink-0">
+    <h3 class="text-sm font-medium text-secondary-700">Document Sections</h3>
+  </div>
+  <div class="sidebar-content p-4 overflow-y-auto flex-grow">
+    <ul class="section-list space-y-2">
+      <li 
+        *ngFor="let section of sections" 
+        class="section-item cursor-pointer text-sm p-1"
+        [class.active]="selectedSection?.id === section.id"
+        (click)="selectSection(section.id)"
+      >
+        <span class="section-title">{{ section.title }}</span>
+      </li>
+    </ul>
+  </div>
+</div>
