+++ ui/src/app/components/core/document-sections/document-sections.component.scss
@@ -1,0 +1,26 @@
+.document-sections {
+  .sidebar-header {
+    background-color: #f9fafb;
+  }
+  
+  .section-list {
+    .section-item {
+      border-radius: 4px;
+      transition: background-color 0.2s ease;
+      
+      &:hover {
+        background-color: rgba(59, 130, 246, 0.05);
+      }
+      
+      &.active {
+        background-color: rgba(59, 130, 246, 0.1);
+        font-weight: 500;
+        color: #2563eb;
+        
+        .section-title {
+          color: #2563eb;
+        }
+      }
+    }
+  }
+}
