--- ui/src/app/components/shared/inline-edit/inline-edit.component.html
@@ -1,126 +1,0 @@
-<!-- Inline edit overlay -->
-<div *ngIf="isActive" class="inline-edit-overlay" (keydown.escape)="cancelEdit()">
-  <!-- Prompt input UI - Copilot style -->
-  <div
-    *ngIf="showPromptInput"
-    class="inline-edit-container"
-    [style.position]="'absolute'"
-    [style.top.px]="positionTop"
-    [style.left.px]="positionLeft"
-  >
-    <div class="p-3 flex flex-col">
-      <div class="flex items-center justify-between mb-2">
-        <div class="text-xs font-medium text-gray-500 flex items-center">
-          <ng-icon
-            name="heroSparklesSolid"
-            class="text-primary-500 mr-1"
-            size="16"
-          ></ng-icon>
-          Edit with AI
-          <span
-            class="text-[7px] bg-yellow-100 text-yellow-700 font-semibold px-2 py-0.5 ml-2 rounded-full uppercase tracking-wide"
-          >
-            Beta
-          </span>
-        </div>
-        <button
-          class="text-gray-400 hover:text-gray-600"
-          (click)="cancelEdit()"
-        >
-          <ng-icon name="heroXMark" size="14"></ng-icon>
-        </button>
-      </div>
-
-      <textarea
-        [(ngModel)]="userPrompt"
-        class="w-full p-2 border border-gray-200 rounded text-xs min-h-[60px] mb-2"
-        placeholder="Enter instructions for AI (e.g., 'Make this more concise', 'Fix grammar', 'Use professional tone')"
-        autofocus
-        (keydown.escape)="cancelEdit()"
-      ></textarea>
-
-      <div class="flex justify-end">
-        <button
-          class="px-2 py-1 bg-primary-500 text-white rounded text-xs font-medium hover:bg-primary-600 flex items-center"
-          (click)="submitPrompt()"
-        >
-          <ng-icon name="heroSparkles" class="mr-1" size="12"></ng-icon>
-          Generate
-        </button>
-      </div>
-    </div>
-  </div>
-
-  <!-- AI suggestion UI - Copilot style -->
-  <div
-    *ngIf="!showPromptInput"
-    class="inline-edit-container"
-    [style.position]="'absolute'"
-    [style.top.px]="positionTop"
-    [style.left.px]="positionLeft"
-  >
-    <div class="p-3 flex flex-col">
-      <div class="flex items-center justify-between mb-2">
-        <div class="text-xs font-medium text-gray-500 flex items-center">
-          <ng-icon
-            name="heroSparklesSolid"
-            class="text-primary-500 mr-1"
-            size="16"
-          ></ng-icon>
-          AI Suggestion
-          <span *ngIf="isLoading" class="ml-2 text-xs text-primary-500">
-            Generating...
-          </span>
-        </div>
-        <button
-          class="text-gray-400 hover:text-gray-600"
-          (click)="cancelEdit()"
-        >
-          <ng-icon name="heroXMark" size="14"></ng-icon>
-        </button>
-      </div>
-
-      <div
-        class="p-2 rounded text-xs min-h-[60px] mb-2 relative"
-        [class.bg-primary-50]="!isLoading"
-        [class.border-primary-100]="!isLoading"
-        [class.border]="!isLoading"
-        [class.bg-gray-50]="isLoading"
-        [class.border-gray-100]="isLoading"
-        [class.border]="isLoading"
-      >
-        <div
-          *ngIf="isLoading"
-          class="absolute inset-0 flex items-center justify-center"
-        >
-          <div class="animate-pulse flex space-x-1">
-            <div class="h-1.5 w-1.5 bg-primary-500 rounded-full"></div>
-            <div class="h-1.5 w-1.5 bg-primary-500 rounded-full"></div>
-            <div class="h-1.5 w-1.5 bg-primary-500 rounded-full"></div>
-          </div>
-        </div>
-        <div
-          [innerHTML]="renderedHtml"
-          class="max-h-[200px] overflow-y-auto"
-        ></div>
-      </div>
-
-      <div class="flex justify-end space-x-2">
-        <button
-          class="px-2 py-1 border border-gray-200 rounded text-xs font-medium text-gray-600 hover:bg-gray-50"
-          (click)="cancelEdit()"
-        >
-          Dismiss
-        </button>
-        <button
-          class="px-2 py-1 bg-primary-500 text-white rounded text-xs font-medium hover:bg-primary-600 flex items-center"
-          [disabled]="isLoading || !editedText"
-          (click)="applyEdit()"
-        >
-          <ng-icon name="heroCheck" class="mr-1" size="12"></ng-icon>
-          Apply
-        </button>
-      </div>
-    </div>
-  </div>
-</div>
