--- ui/src/app/components/shared/inline-edit/inline-edit.component.scss
@@ -1,153 +1,0 @@
-.inline-edit-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  z-index: 1000;
-  pointer-events: none;
-}
-
-.inline-edit-container {
-  background-color: white;
-  border-radius: 6px;
-  box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.15);
-  width: 360px;
-  pointer-events: auto;
-}
-
-/* 
-  Global style for the highlight - this will be applied to the parent document
-  The :global pseudo-selector allows the style to be applied outside the component
-*/
-/* Standard inline selection */
-:global(.inline-edit-selection) {
-  background-color: rgba(59, 130, 246, 0.3) !important; /* Light blue highlight - slightly more visible */
-  border-radius: 2px !important;
-  padding: 1px 0 !important;
-  border-bottom: 1px dotted rgba(59, 130, 246, 0.7) !important; /* Add a subtle dotted underline for better visibility */
-  box-shadow: 0 0 0 1px rgba(59, 130, 246, 0.1) !important; /* Add a very subtle glow */
-  position: relative !important;
-  z-index: 10 !important; /* Ensure the highlight appears below the inline edit UI but above other content */
-  pointer-events: none !important; /* Allow clicking through the highlight */
-}
-
-/* Node level highlighting for complex selections */
-:global(.inline-edit-selection-node) {
-  background-color: rgba(59, 130, 246, 0.35) !important;
-  border-radius: 1px !important;
-  border-bottom: 1px dotted rgba(59, 130, 246, 0.8) !important;
-  box-shadow: 0 0 0 1px rgba(59, 130, 246, 0.15) !important;
-  display: inline !important;
-  white-space: pre-wrap !important; /* Preserve whitespace */
-}
-
-/* Persistent highlighting that remains during editing */
-:global(.inline-edit-selection-persistent) {
-  background-color: rgba(59, 130, 246, 0.35) !important;
-  border-radius: 3px !important;
-  padding: 2px !important;
-  border-bottom: 2px dotted rgba(59, 130, 246, 0.8) !important;
-  box-shadow: 0 0 0 2px rgba(59, 130, 246, 0.15) !important;
-  position: relative !important;
-  z-index: 20 !important;
-  display: inline-block !important;
-}
-
-/* Style for the editor when inline editing is active */
-:global(.inline-edit-active) {
-  position: relative !important;
-}
-
-:global(.inline-edit-active::after) {
-  content: '' !important;
-  position: absolute !important;
-  top: 0 !important;
-  left: 0 !important;
-  right: 0 !important;
-  bottom: 0 !important;
-  border: 2px solid rgba(59, 130, 246, 0.3) !important;
-  pointer-events: none !important;
-  z-index: 5 !important;
-}
-
-/* Overlay container and elements */
-:global(.selection-overlay-container) {
-  position: absolute !important;
-  top: 0 !important;
-  left: 0 !important;
-  right: 0 !important;
-  bottom: 0 !important;
-  pointer-events: none !important;
-  z-index: 1000 !important;
-}
-
-:global(.inline-edit-selection-overlay) {
-  background-color: rgba(59, 130, 246, 0.3) !important;
-  border-radius: 2px !important;
-  border-bottom: 1px dotted rgba(59, 130, 246, 0.7) !important;
-  position: absolute !important;
-  box-shadow: 0 0 0 2px rgba(59, 130, 246, 0.1) !important;
-  pointer-events: none !important;
-  z-index: 50 !important;
-}
-
-@keyframes slideUp {
-  from { transform: translateX(-50%) translateY(10px); opacity: 0; }
-  to { transform: translateX(-50%) translateY(0); opacity: 1; }
-}
-
-:host ::ng-deep {
-  .max-h-200px {
-    max-height: 200px;
-  }
-  
-  // Style for code blocks in the edited text
-  pre {
-    background-color: #f8f9fa;
-    border-radius: 0.25rem;
-    padding: 0.5rem;
-    margin: 0.5rem 0;
-    overflow-x: auto;
-    font-size: 0.75rem;
-    line-height: 1.25;
-  }
-  
-  // Style for inline code in the edited text
-  code {
-    background-color: #f8f9fa;
-    border-radius: 0.25rem;
-    padding: 0.125rem 0.25rem;
-    font-size: 0.75rem;
-  }
-  
-  // Style for lists in the edited text
-  ul, ol {
-    padding-left: 1.5rem;
-    margin: 0.5rem 0;
-  }
-  
-  ul li {
-    list-style-type: disc;
-  }
-  
-  ol li {
-    list-style-type: decimal;
-  }
-  
-  // Style for links in the edited text
-  a {
-    color: #3b82f6;
-    text-decoration: underline;
-  }
-  
-  // Style for emphasis in the edited text
-  em {
-    font-style: italic;
-  }
-  
-  // Style for strong emphasis in the edited text
-  strong {
-    font-weight: 600;
-  }
-}
